requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
blessings==1.7
bpython==0.17.1
celery==3.1.26.post2
certifi==2018.11.29
chardet==3.0.4
curtsies==0.3.0
Django==2.1.2
django-admin==1.3.2
django-celery==3.2.2
django-cors-headers==2.5.2
django-debug-toolbar==1.11
django-excel-response2==2.0.8
django-extensions==2.1.3
django-filter==2.0.0
django-js-asset==1.1.0
django-mptt==0.9.1
django-render-block==0.5
django-sendgrid-v5==0.7.0
django-six==1.0.4
django-templated-email==2.2.0
djangorestframework==3.8.2
et-xmlfile==1.0.1
future==0.17.1
greenlet==0.4.15
idna==2.8
image==1.5.27
jdcal==1.4
kombu==3.0.37
mysqlclient==1.3.13
openpyxl==2.5.12
Pillow==5.4.1
Pygments==2.3.1
python-http-client==3.1.0
pytz==2018.9
redis==2.10.6
requests==2.21.0
screen==1.0.1
sendgrid==5.6.0
six==1.12.0
sqlparse==0.2.4
typing==3.6.6
urllib3==1.24.1
vine==1.2.0
wcwidth==0.1.7
xlwt==1.3.0
